Fix three type conversion warnings
- Add explicit NSInteger casts in PBGitHistoryController.m:300,301 for NSUInteger to NSInteger conversion when comparing and accessing array counts
- Add explicit NSInteger cast in PBGitHistoryController.m:328 for NSUInteger to NSInteger conversion from firstIndex
- Add explicit NSInteger casts in OpenRecentController.m:108,120 for NSUInteger to NSInteger conversion in scrollRowToVisible calls

These fixes resolve implicit sign conversion warnings while maintaining original logic.

+++ Classes/Controllers/OpenRecentController.m
@@ -105,7 +105,7 @@
 			NSUInteger index = (foundIndex == NSNotFound || foundIndex == 0) ? 0 : foundIndex - 1;
 			selectedResult = [currentResults objectAtIndex:index];
 			[resultViewer selectRowIndexes:[NSIndexSet indexSetWithIndex:index] byExtendingSelection:FALSE];
-			[resultViewer scrollRowToVisible:index];
+			[resultViewer scrollRowToVisible:(NSInteger)index];
 		}
 		result = YES;
 	}
@@ -117,7 +117,7 @@
 			if(index >= count && count > 0) index = count - 1;
 			selectedResult = [currentResults objectAtIndex:index];
 			[resultViewer selectRowIndexes:[NSIndexSet indexSetWithIndex:index] byExtendingSelection:FALSE];
-			[resultViewer scrollRowToVisible:index];
+			[resultViewer scrollRowToVisible:(NSInteger)index];
 		}
 		result = YES;
 	}
